Fix sampling corner-case for half-transition ERs.
When storing (s, a, r, d) transitions instead of full ones we need to
look ahead in the ER buffer and use the state at the next step as our
next_state. This manages better the situation in which we sample from
the last position in the ER, which happens frequently in the case of
prioritized replay.